Add long press and release event support
This adds long press support and basic for release
event support. This is done by overriding @final
methods of homeassistant's button for which it is
unclear why they are marked as such.

Attributes for type of click are now set to "long"
for a long press and "single" for a short press.